energy_forecast/api/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
xgboost==1.4.2
python-multipart==0.0.5
email-validator==1.1.3
requests==2.26.0
aiohttp==3.7.4
python-json-logger==2.0.2
prometheus-client==0.11.0
python-decouple==3.4
